Move custom message box controls to new namespace
