+++ nvQuickSite/Controls/Dialogs/MsgBoxOk.resx
