+++ nvQuickSite/Controls/Dialogs/MsgBoxYesNo.resx
