+++ nvQuickSite/Controls/Dialogs/MsgBoxYesNoIgnore.resx
